[TEST] Service unit tests

+++ eclipse-workspace/dailyorg_java/src/main/java/fr/nexa/dailyorg/service/dailyorg/impl/OrganizerUserService.java
@@ -1,14 +1,15 @@
 package fr.nexa.dailyorg.service.dailyorg.impl;
 
 import java.util.Optional;
+import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import fr.nexa.dailyorg.model.AppUser;
 import fr.nexa.dailyorg.model.dailyorg.OrganizerUser;
 import fr.nexa.dailyorg.repository.dailyorg.IOrganizerUserRepository;
 import fr.nexa.dailyorg.service.dailyorg.IOrganizerUserService;
+import fr.nexa.dailyorg.utils.EErrorMessages;
 import lombok.AllArgsConstructor;
 
 @Service
@@ -18,27 +19,37 @@ public class OrganizerUserService implements IOrganizerUserService{
 	private final IOrganizerUserRepository organizerUserRepository;
 	
 	@Override
-	public OrganizerUser findByAppUserId(AppUser userId) {
-		return organizerUserRepository.findByAppUser(userId);
-	}
-	
-	@Override
-	public Optional<OrganizerUser> findByAppUserId(long userId) {
+	public Optional<OrganizerUser> findByAppUserId(Long userId) {
+		if(userId == null) {
+			return Optional.empty();
+		}
 		return organizerUserRepository.findById(userId);
 	}
 	
 	@Override
 	public OrganizerUser create(OrganizerUser organizerUser) {
+		if(organizerUser == null) {
+			Logger.getLogger(this.getClass().getName()).severe(EErrorMessages.NULL_VALUE.getMessage());
+			throw new IllegalArgumentException(EErrorMessages.NULL_VALUE.getMessage());
+		}
 		return organizerUserRepository.save(organizerUser);
 	}
 	
 	@Override
 	public void delete(OrganizerUser organizerUser) {
+		if(organizerUser == null) {
+			Logger.getLogger(this.getClass().getName()).severe(EErrorMessages.NULL_VALUE.getMessage());
+			throw new IllegalArgumentException(EErrorMessages.NULL_VALUE.getMessage());
+		}
 		organizerUserRepository.delete(organizerUser);
 	}
 	
 	@Override
 	public OrganizerUser update(OrganizerUser organizerUser) {
+		if(organizerUser == null) {
+			Logger.getLogger(this.getClass().getName()).severe(EErrorMessages.NULL_VALUE.getMessage());
+			throw new IllegalArgumentException(EErrorMessages.NULL_VALUE.getMessage());
+		}
 		return organizerUserRepository.save(organizerUser);
 	}
